+++ scripts/compute_language_breakdown.py
@@ -3,39 +3,8 @@ import pandas as pd
 import json
 import argparse
 
-from sklearn.metrics import roc_auc_score
-
-
-def compute_auc(y_true, y_pred):
-    try:
-        return roc_auc_score(y_true, y_pred)
-    except ValueError:
-        return np.nan
-
-
-def compute_subgroup_auc(df, subgroup, label, model_name):
-    subgroup_examples = df[df[subgroup]]
-    return compute_auc(subgroup_examples[label], subgroup_examples[model_name])
-
-
-def convert_dataframe_to_bool(df):
-    bool_df = df.copy()
-    for la in LANGS:
-        bool_df[la] = df.lang == la
-    return bool_df
-
-
-def compute_lang_metrics_for_model(dataset, subgroups, model, label_col):
-    """Computes per-subgroup metrics for all subgroups and one model."""
-    records = []
-    for subgroup in subgroups:
-        record = {
-            "subgroup": subgroup,
-            "subgroup_size": len(dataset[dataset[subgroup]]),
-        }
-        record[SUBGROUP_AUC] = compute_subgroup_auc(dataset, subgroup, label_col, model)
-        records.append(record)
-    return pd.DataFrame(records).sort_values("subgroup_auc", ascending=True)
+from detoxify.bias_metrics import convert_dataframe_to_bool, compute_lang_metrics_for_model, \
+    MODEL_NAME, LANGS, TOXICITY_COLUMN
 
 
 def main():
@@ -66,11 +35,5 @@ if __name__ == "__main__":
     )
     args = parser.parse_args()
     TEST = args.res_path
-
-    LANGS = ["es", "fr", "pt", "ru", "it", "tr"]
-
-    TOXICITY_COLUMN = "toxic"
-    MODEL_NAME = "XLM-R"
-    SUBGROUP_AUC = "subgroup_auc"
 
     main()
